Fix Shared Library
Note that the linker will only generate a .lib file if a symbol is properly decorated with
an export directive.

So for the library defining the exported symbol is must be
__declspec(dllexport).
And for the project consuming the symbol the directive must be:
__declspec(dllimport)

CMake can handle this, even going so far as to generate a header file
complete with all the macro's you would ever want.

So this commit makes it so that everything will compile nicely since the
SharedLib.lib file gets created.

But it still fails at runtime since the SharedLib.dll still needs to get copied to the output
binary directory.

I have to figure that out next.